backend/requirements.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
annotated-types==0.6.0
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cookiecutter==2.5.0
cymem==2.0.8
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==5.0.1
django-admin-rangefilter==0.12.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-import-export==3.3.5
django-tailwind==3.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.13.1
fsspec==2024.3.1
gunicorn==21.2.0
httpie==3.2.2
idna==3.6
Jinja2==3.1.3
joblib==1.4.0
langcodes==3.3.0
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.4
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
odfpy==1.4.1
openai-whisper==20231117
openpyxl==3.1.2
packaging==23.2
platformdirs==4.2.0
preshed==3.0.9
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
setuptools==69.1.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
sympy==1.12
tablib==3.5.0
text-unidecode==1.3
textblob==0.18.0.post0
thinc==8.2.3
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
typer==0.9.4
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
wasabi==1.1.2
weasel==0.3.4
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
